requirements.txt
blinker==1.8.2
branca==0.7.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
convertdate==2.4.0
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
ephem==4.1.5
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
folium==0.17.0
fonttools==4.53.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
hijri-converter==2.3.1
holidays==0.25
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==1.26.4
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
pillow==10.4.0
plotly==5.22.0
prophet==1.1.5
pycparser==2.22
PyJWT==2.7.0
PyMeeus==0.5.12
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools-git==1.2
six==1.16.0
SQLAlchemy==2.0.31
stanio==0.5.1
tenacity==8.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
xyzservices==2024.6.0
zipp==3.19.2
ortools